// server/vance-addon-brain-finance/client/src/generated/finance/FinanceNodeDto.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.finance.FinanceNodeDto
*/

import { FinanceValueDto } from './FinanceValueDto';

export interface FinanceNodeDto {
  name: string;
  title?: string;
  icon?: string;
  color?: string;
  sign: number;
  description?: string;
  notesRef?: string;
  values: FinanceValueDto[];
  children: FinanceNodeDto[];
}