// server/vance-addon-brain-finance/client/src/generated/finance/FinanceTreeDto.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.finance.FinanceTreeDto
*/

import { FinanceNodeDto } from './FinanceNodeDto';

export interface FinanceTreeDto {
  version: number;
  title?: string;
  description?: string;
  root?: FinanceNodeDto;
}